+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## 2.2.4 (...)
 - introduce `userIdAttribute` for `FormLdapAuthentication` to "normalize" the
   user ID used inside the VPN service (issue #151)
+- implement `addRealm` option for `FormLdapAuthentication` that adds a provided
+  domain to the user specified "authUser" if no domain is specified yet
 
 ## 2.2.3 (2020-03-23)
 - add German portal translation